+++ src/main/java/com/sparksInTheStep/webBoard/post/presentation/dto/PostRequest.java
@@ -1,4 +1,4 @@
-package com.sparksInTheStep.webBoard.post.controller.dto;
+package com.sparksInTheStep.webBoard.post.presentation.dto;
 
 
 import jakarta.validation.constraints.NotBlank;
